refactor: MovieList 컴포넌트 분리
MovieList.tsx
- 전달받은 movies 기반으로 검색,필터링
- 화면에 props로 받은 영화 리스트 렌더링
